+++ jsonschema_restructuredtext/__init__.py
@@ -1,3 +1,3 @@
 from jsonschema_restructuredtext.converter.rst import generate
 
-generate = generate
+__all__ = ["generate"]