+++ jsonschema_restructuredtext/constants.py
@@ -1,2 +1,2 @@
 # Define default section level punctuation
-DEFAULT_SECTION_PUNCTUATION = ["=", "-", "^", "~", '+', '*', '+', '.']
+DEFAULT_SECTION_PUNCTUATION = ["=", "-", "^", "~", "+", "*", "#", "."]